Level increases, bugfixes, and Powerups
Fixed many, many bugs.
Player can only shoot 2 bullets initially.
Added a way for enemies to increase in rows when all the enemies in the current level are killed.
Enemies move in a sideways manner.
Added a mechanism for powerups.
Added one powerup, IncreasedNumberOfBullets which increases the number of max bullets the player can fire.
There is a 1/101 chance for a powerup to spawn when an enemy is killed.